+++ src/spyglass/common/common_nwbfile.py
@@ -2,6 +2,7 @@ import os
 import random
 import stat
 import string
+from pathlib import Path
 
 import datajoint as dj
 import numpy as np
@@ -9,9 +10,9 @@ import pandas as pd
 import pynwb
 import spikeinterface as si
 from hdmf.common import DynamicTable
-from pathlib import Path
 
-from ..settings import raw_dir
+from spyglass.settings import analysis_dir, raw_dir
+from spyglass.utils.dj_mixin import SpyglassMixin
 from ..utils.dj_helper_fn import get_child_tables
 from ..utils.nwb_helper_fn import get_electrode_indices, get_nwb_file
 
@@ -38,7 +39,7 @@ NWB_KEEP_FIELDS = (
 
 
 @schema
-class Nwbfile(dj.Manual):
+class Nwbfile(SpyglassMixin, dj.Manual):
     definition = """
     # Table for holding the NWB files.
     nwb_file_name: varchar(64)   # name of the NWB file
@@ -147,7 +148,7 @@ class Nwbfile(dj.Manual):
 # TODO: add_to_kachery will not work because we can't update the entry after it's been used in another table.
 # We therefore need another way to keep track of the
 @schema
-class AnalysisNwbfile(dj.Manual):
+class AnalysisNwbfile(SpyglassMixin, dj.Manual):
     definition = """
     # Table for holding the NWB files that contain results of analysis, such as spike sorting.
     analysis_file_name: varchar(64)               # name of the file
@@ -296,7 +297,7 @@ class AnalysisNwbfile(dj.Manual):
     def get_abs_path(analysis_nwb_file_name):
         """Return the absolute path for a stored analysis NWB file given just the file name.
 
-        The SPYGLASS_BASE_DIR environment variable must be set.
+        The spyglass config from settings.py must be set.
 
         Parameters
         ----------
@@ -308,25 +309,16 @@ class AnalysisNwbfile(dj.Manual):
         analysis_nwb_file_abspath : str
             The absolute path for the given file name.
         """
-        base_dir = Path(os.getenv("SPYGLASS_BASE_DIR", None))
-        assert (
-            base_dir is not None
-        ), "You must set SPYGLASS_BASE_DIR environment variable."
-
         # see if the file exists and is stored in the base analysis dir
-        test_path = str(base_dir / "analysis" / analysis_nwb_file_name)
+        test_path = f"{analysis_dir}/{analysis_nwb_file_name}"
 
         if os.path.exists(test_path):
             return test_path
         else:
             # use the new path
-            analysis_file_base_path = (
-                base_dir
-                / "analysis"
-                / AnalysisNwbfile.__get_analysis_file_dir(
-                    analysis_nwb_file_name
-                )
-            )
+            analysis_file_base_path = Path(
+                analysis_dir
+            ) / AnalysisNwbfile.__get_analysis_file_dir(analysis_nwb_file_name)
             if not analysis_file_base_path.exists():
                 os.mkdir(str(analysis_file_base_path))
             return str(analysis_file_base_path / analysis_nwb_file_name)
@@ -649,7 +641,7 @@ class AnalysisNwbfile(dj.Manual):
 
 
 @schema
-class NwbfileKachery(dj.Computed):
+class NwbfileKachery(SpyglassMixin, dj.Computed):
     definition = """
     -> Nwbfile
     ---
@@ -667,7 +659,7 @@ class NwbfileKachery(dj.Computed):
 
 
 @schema
-class AnalysisNwbfileKachery(dj.Computed):
+class AnalysisNwbfileKachery(SpyglassMixin, dj.Computed):
     definition = """
     -> AnalysisNwbfile
     ---