+++ src/spyglass/lfp/lfp_merge.py
@@ -7,33 +7,34 @@ from spyglass.common.common_interval import IntervalList  # noqa: F401
 from spyglass.lfp.lfp_imported import ImportedLFP  # noqa: F401
 from spyglass.lfp.v1.lfp import LFPV1  # noqa: F401
 from spyglass.utils.dj_merge_tables import _Merge
+from spyglass.utils.dj_mixin import SpyglassMixin
 
 schema = dj.schema("lfp_merge")
 
 
 @schema
-class LFPOutput(_Merge):
+class LFPOutput(_Merge, SpyglassMixin):
     definition = """
     merge_id: uuid
     ---
     source: varchar(32)
     """
 
-    class LFPV1(dj.Part):  # noqa: F811
+    class LFPV1(SpyglassMixin, dj.Part):  # noqa: F811
         definition = """
         -> master
         ---
         -> LFPV1
         """
 
-    class ImportedLFP(dj.Part):  # noqa: F811
+    class ImportedLFP(SpyglassMixin, dj.Part):  # noqa: F811
         definition = """
         -> master
         ---
         -> ImportedLFP
         """
 
-    class CommonLFP(dj.Part):  # noqa: F811
+    class CommonLFP(SpyglassMixin, dj.Part):  # noqa: F811
         """Table to pass-through legacy LFP"""
 
         definition = """