+++ src/spyglass/spikesorting/curation_figurl.py
@@ -1,16 +1,16 @@
-import datajoint as dj
+from typing import Any, Dict, List, Union
 
-from typing import Any, Union, List, Dict
+import datajoint as dj
+import kachery_cloud as kcl
+import sortingview.views as vv
+import spikeinterface as si
+from sortingview.SpikeSortingView import SpikeSortingView
+
+from spyglass.utils.dj_mixin import SpyglassMixin
 
 from .spikesorting_curation import Curation
 from .spikesorting_recording import SpikeSortingRecording
 from .spikesorting_sorting import SpikeSorting
-
-import spikeinterface as si
-
-from sortingview.SpikeSortingView import SpikeSortingView
-import kachery_cloud as kcl
-import sortingview.views as vv
 
 schema = dj.schema("spikesorting_curation_figurl")
 
@@ -25,7 +25,7 @@ schema = dj.schema("spikesorting_curation_figurl")
 
 
 @schema
-class CurationFigurlSelection(dj.Manual):
+class CurationFigurlSelection(SpyglassMixin, dj.Manual):
     definition = """
     -> Curation
     ---
@@ -34,7 +34,7 @@ class CurationFigurlSelection(dj.Manual):
 
 
 @schema
-class CurationFigurl(dj.Computed):
+class CurationFigurl(SpyglassMixin, dj.Computed):
     definition = """
     -> CurationFigurlSelection
     ---